fix(core): commandPrefix word boundary and compound command safety
This change addresses two security/correctness issues in shell policy:
1. strict word boundaries for commandPrefix (e.g., 'git log' no longer matches 'git logout')
2. recursive safety checks for compound commands (e.g., 'git log && rm -rf /' is now properly vetted)
